+++ NEWS.md
@@ -1,5 +1,10 @@
 NEWS for the htmlTable package
 
+Changes fro 2.1.0
+-----------------
+* Added options for how the empty spacer cell appears (see `addHtmlTableStyle()` spacer.* options)
+* Fixed ignored `align.header`
+
 Changes for 2.0.1
 -----------------
 * Fix for txtRound not handling negative numbers (issue #76)